+++ app/src/main/java/com/ssafy/neegongnaegong/presentation/calendar/component/calendar/CalendarCell.kt
@@ -22,6 +22,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.ssafy.neegongnaegong.presentation.ui.theme.NeeGongNaeGongTheme
+import com.ssafy.neegongnaegong.presentation.util.color
 import java.time.DayOfWeek
 import java.time.LocalDate
