+++ app/src/main/java/com/ssafy/neegongnaegong/presentation/component/picker/date/DatePickerBody.kt
@@ -3,8 +3,6 @@ package com.ssafy.neegongnaegong.presentation.component.picker.date
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
@@ -15,7 +13,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.ssafy.neegongnaegong.presentation.calendar.component.calendar.DayOfWeek
 import com.ssafy.neegongnaegong.presentation.ui.theme.NeeGongNaeGongTheme
 import java.time.LocalDate
 import java.time.YearMonth
@@ -32,8 +29,6 @@ fun DatePickerBody(
     val firstDayOfWeek by remember { mutableIntStateOf(selectedMonth.atDay(1).dayOfWeek.value % 7) }
 
     Column(modifier = modifier) {
-        DayOfWeek()
-        Spacer(Modifier.height(16.dp))
         Column {
             val totalRows = (firstDayOfWeek + lastDay + 6) / 7
             for (row in 0 until totalRows) {
